[develop] 싱글톤 클래스 ImageProcessor 구현
